+++ test/test_facebook.py
@@ -16,6 +16,7 @@
 import facebook
 import os
 import unittest
+import inspect
 
 try:
     from urllib.parse import parse_qs, urlencode, urlparse
@@ -25,7 +26,11 @@ except ImportError:
 
 
 class FacebookTestCase(unittest.TestCase):
-    """Sets up application ID and secret from environment."""
+    """
+    Sets up application ID and secret from environment and initialises an
+    empty list for test users.
+
+    """
     def setUp(self):
         try:
             self.app_id = os.environ["FACEBOOK_APP_ID"]
@@ -33,6 +38,18 @@ class FacebookTestCase(unittest.TestCase):
         except KeyError:
             raise Exception("FACEBOOK_APP_ID and FACEBOOK_SECRET "
                             "must be set as environmental variables.")
+
+        self.test_users = []
+
+    def tearDown(self):
+        """Deletes the test users included in the test user list."""
+        token = facebook.GraphAPI().get_app_access_token(
+            self.app_id, self.secret)
+        graph = facebook.GraphAPI(token)
+
+        for user in self.test_users:
+            graph.request(user['id'], {}, None, method='DELETE')
+        del self.test_users[:]
 
     def assert_raises_multi_regex(
             self, expected_exception, expected_regexp, callable_obj=None,
@@ -47,6 +64,26 @@ class FacebookTestCase(unittest.TestCase):
             callable_obj(*args)
         except facebook.GraphAPIError as error:
             self.assertEqual(error.message, expected_regexp)
+
+    def create_test_users(self, app_id, graph, amount):
+        """Function for creating test users."""
+        for i in range(amount):
+            u = graph.request(app_id + '/accounts/test-users', {}, {},
+                              method='POST')
+            self.test_users.append(u)
+
+    def create_friend_connections(self, user, friends):
+        """Function for creating friend connections for a test user."""
+        user_graph = facebook.GraphAPI(user['access_token'])
+
+        for friend in friends:
+            if user['id'] == friend['id']:
+                continue
+            user_graph.request(user['id'] + '/friends/' + friend['id'],
+                               {}, {}, method='POST')
+            respondent_graph = facebook.GraphAPI(friend['access_token'])
+            respondent_graph.request(friend['id'] + '/friends/' + user['id'],
+                                     {}, {}, method='POST')
 
 
 class TestGetAppAccessToken(FacebookTestCase):
@@ -195,6 +232,40 @@ class TestParseSignedRequest(FacebookTestCase):
         self.assertTrue('code' in result)
         self.assertTrue('user_id' in result)
         self.assertTrue('algorithm' in result)
+
+
+class TestGetAllConnectionsMethod(FacebookTestCase):
+
+    def test_function_with_zero_connections(self):
+        token = facebook.GraphAPI().get_app_access_token(
+            self.app_id, self.secret)
+        graph = facebook.GraphAPI(token)
+
+        self.create_test_users(self.app_id, graph, 1)
+        friends = graph.get_all_connections(self.test_users[0]['id'],
+                                            'friends')
+
+        self.assertTrue(inspect.isgenerator(friends))
+        self.assertTrue(len(list(friends)) == 0)
+
+    def test_function_returns_correct_connections(self):
+        token = facebook.GraphAPI().get_app_access_token(
+            self.app_id, self.secret)
+        graph = facebook.GraphAPI(token)
+
+        self.create_test_users(self.app_id, graph, 27)
+        self.create_friend_connections(self.test_users[0], self.test_users)
+
+        friends = graph.get_all_connections(self.test_users[0]['id'],
+                                            'friends')
+        self.assertTrue(inspect.isgenerator(friends))
+
+        friends_list = list(friends)
+        self.assertTrue(len(friends_list) == 26)
+        for f in friends:
+            self.assertTrue(isinstance(f, dict))
+            self.assertTrue('name' in f)
+            self.assertTrue('id' in f)
 
 
 if __name__ == '__main__':