+++ docs/changes.rst
@@ -8,7 +8,9 @@ Version 3.0.0 (unreleased)
  - Add support for Graph API version 2.8.
  - Remove support for Graph API version 2.1.
  - Change default Graph API version to 2.2.
- - Add support for requests' sessions.
+ - Add support for requests' sessions (#201).
+ - Add versioning to access token endpoints (#322).
+ - Add new `get_all_connections` method to make pagination easier (#337).
 
 Version 2.0.0 (2016-08-08)
 ==============================